requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
PyMySQL>=1.0
python-dotenv>=1.0
Flask-JWT-Extended>=4.4
Jinja2>=3.1
passlib[bcrypt]>=1.7
